Barbershop edit

+++ src/main/java/org/example/repository/BarbershopRepository.java
@@ -3,6 +3,6 @@ package org.example.repository;
 import org.example.entity.Barbershop;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BarbershopRepository extends JpaRepository<Barbershop, Long> {
-    Barbershop findBarbershopById(Long id);
+public interface BarbershopRepository extends JpaRepository<Barbershop, Integer> {
+    Barbershop findBarbershopById(Integer id);
 }
